ui and settings updates

+++ src/Components/Forms/DeleteDataHolderForm.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import "../../Styles/Components/form.css";
+import { notify } from "../Notification/Notification";
 
 /**
  * Importing actions here
@@ -29,10 +30,17 @@ class DeleteDataHolderForm extends Component {
 
   handleSubmit = (e) => {
     e.preventDefault();
-    if (this.state.DHId === this.props.dataholder.id) {
-      this.props.deleteDataHolder(this.state);
-      this.props.openModal();
+    if (this.props.uid !== "9ALPoJY04RRCBqGnHzeB0qU1FzJ3") {
+      if (this.state.DHId === this.props.dataholder.id) {
+        this.props.deleteDataHolder(this.state);
+      } else {
+        notify("Entered ID is worng. Please, try again.");
+      }
+    } else {
+      notify("You do not have enough permission to perform this action.");
     }
+
+    this.props.openModal();
   };
 
   render() {
@@ -90,6 +98,13 @@ class DeleteDataHolderForm extends Component {
   }
 }
 
+const mapStateToProps = (state) => {
+  const uid = state.firebase.auth.uid;
+  return {
+    uid: uid,
+  };
+};
+
 const mapDispatchToProps = (dispatch) => {
   return {
     deleteDataHolder: (dataholder) => {
@@ -98,4 +113,7 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(null, mapDispatchToProps)(DeleteDataHolderForm);
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(DeleteDataHolderForm);
